Move isCommentsTreeInitialized
