+++ client/state/comments/selectors/is-comments-tree-initialized.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { get } from 'lodash';
-
-/**
  * Internal dependencies
  */
 import 'state/comments/init';
@@ -16,11 +11,11 @@ import 'state/comments/init';
  * @param  {string}  status   unapproved|approved|all|spam|trash
  * @returns {boolean} True if the comment tree has been initialized
  */
-export default function isCommentsTreeInitialized( state, siteId, status ) {
+export function isCommentsTreeInitialized( state, siteId, status ) {
 	if ( 'all' === status ) {
-		const tree = get( state, [ 'comments', 'treesInitialized', siteId ] );
+		const tree = state.comments.treesInitialized[ siteId ];
 		return tree && ( tree.approved || tree.unapproved );
 	}
 
-	return get( state, [ 'comments', 'treesInitialized', siteId, status ], false );
+	return state.comments.treesInitialized[ siteId ]?.[ status ] ?? false;
 }
